+++ geofileops/util/_geoops_sql.py
@@ -52,7 +52,7 @@ def buffer(
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
@@ -115,7 +115,7 @@ def convexhull(
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = False,  # Should become True
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
@@ -163,7 +163,7 @@ def delete_duplicate_geometries(
     output_layer: Optional[str] = None,
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     force: bool = False,
 ):
@@ -284,9 +284,9 @@ def makevalid(
     output_layer: Optional[str] = None,
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
-    force_output_geometrytype: Optional[GeometryType] = None,
+    force_output_geometrytype: Union[str, None, GeometryType] = None,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
@@ -305,7 +305,11 @@ def makevalid(
         input_layerinfo = gfo.get_layerinfo(input_path, input_layer)
         force_output_geometrytype = input_layerinfo.geometrytype
         if not explodecollections:
+            assert isinstance(force_output_geometrytype, GeometryType)
             force_output_geometrytype = force_output_geometrytype.to_multitype
+    if isinstance(force_output_geometrytype, str):
+        force_output_geometrytype = GeometryType[force_output_geometrytype]
+    assert force_output_geometrytype is not None
 
     # Init + prepare sql template for this operation
     # ----------------------------------------------
@@ -370,7 +374,7 @@ def select(
     explodecollections: bool = False,
     force_output_geometrytype: Optional[GeometryType] = None,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     nb_parallel: int = 1,
     batchsize: int = -1,
     force: bool = False,
@@ -428,7 +432,7 @@ def simplify(
     columns: Optional[List[str]] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     nb_parallel: int = -1,
     batchsize: int = -1,
@@ -632,6 +636,7 @@ def _single_layer_vector_operation(
 
         # Add application of gridsize around sql_template if specified
         if geom_selected and gridsize != 0.0:
+            assert force_output_geometrytype is not None
             gridsize_op = _format_apply_gridsize_operation(
                 geometrycolumn=f"sub_gridsize.{input_layerinfo.geometrycolumn}",
                 gridsize=gridsize,
@@ -1113,8 +1118,8 @@ def erase(
     #   distinction whether the subquery is finding a row (no match with spatial index)
     #   or if the difference results in an empty/NULL geometry.
     #   Tried to return EMPTY GEOMETRY from GFO_Difference_Collection, but it didn't
-    #   work to use spatialite's ST_IsEmpty(geom) = 0 to filter on this for an unclear
-    #   reason.
+    #   work to use spatialite's ST_IsEmpty(geom) = 0 to filter on this, probably
+    #   because ST_GeomFromWKB doesn't seem to support empty polygons.
     # - ST_difference(geometry , NULL) gives NULL as result -> handle explicitly
     input1_layer_rtree = "rtree_{input1_layer}_{input1_geometrycolumn}"
     input2_layer_rtree = "rtree_{input2_layer}_{input2_geometrycolumn}"
@@ -1173,6 +1178,7 @@ def erase(
           )
          WHERE geom IS NOT NULL
            AND geom <> 'DIFF_EMPTY'
+           AND ST_IsEmpty(geom) = 0
     """
 
     # Go!
@@ -1572,13 +1578,15 @@ def join_by_location(
               SELECT sub_filter.*
                     {area_inters_column_expression}
                 FROM (
-                  SELECT layer1.{{input1_geometrycolumn}} as geom
+                  SELECT layer1.{{input1_geometrycolumn}} AS geom
                         ,layer1.fid l1_fid
-                        ,layer2.{{input2_geometrycolumn}} as l2_geom
+                        ,layer2.{{input2_geometrycolumn}} AS l2_geom
                         {{layer1_columns_prefix_alias_str}}
                         {{layer2_columns_prefix_alias_str}}
-                        ,ST_relate(layer1.{{input1_geometrycolumn}},
-                                   layer2.{{input2_geometrycolumn}}) as spatial_relation
+                        ,ST_relate(
+                            layer1.{{input1_geometrycolumn}},
+                            layer2.{{input2_geometrycolumn}}
+                         ) AS "GFO_$TEMP$_SPATIAL_RELATION"
                     FROM {{input1_databasename}}."{{input1_layer}}" layer1
                     JOIN {{input1_databasename}}."{input1_layer_rtree}" layer1tree
                       ON layer1.fid = layer1tree.id
@@ -1594,7 +1602,7 @@ def join_by_location(
                    LIMIT -1 OFFSET 0
                   ) sub_filter
                WHERE {spatial_relations_filter.format(
-                    spatial_relation="sub_filter.spatial_relation")}
+                    spatial_relation='sub_filter."GFO_$TEMP$_SPATIAL_RELATION"')}
                LIMIT -1 OFFSET 0
               ) sub_area
            {area_inters_filter}
@@ -1602,7 +1610,7 @@ def join_by_location(
         SELECT sub.geom
               {{layer1_columns_from_subselect_str}}
               {{layer2_columns_from_subselect_str}}
-              ,sub.spatial_relation
+              ,sub."GFO_$TEMP$_SPATIAL_RELATION" AS spatial_relation
               {area_inters_column_in_output}
           FROM layer1_relations_filtered sub
     """
@@ -1616,7 +1624,7 @@ def join_by_location(
             SELECT layer1.{{input1_geometrycolumn}} as geom
                   {{layer1_columns_prefix_alias_str}}
                   {{layer2_columns_prefix_alias_null_str}}
-                  ,NULL as spatial_relation
+                  ,NULL AS spatial_relation
                   {area_inters_column_0_in_output}
               FROM {{input1_databasename}}."{{input1_layer}}" layer1
              WHERE 1=1
@@ -1719,8 +1727,8 @@ def join_nearest(
     input2_path: Path,
     output_path: Path,
     nb_nearest: int,
-    distance: float,
-    expand: bool,
+    distance: Optional[float],
+    expand: Optional[bool],
     input1_layer: Optional[str] = None,
     input1_columns: Optional[List[str]] = None,
     input1_columns_prefix: str = "l1_",
@@ -2530,8 +2538,10 @@ def _two_layer_vector_operation(
             if SPATIALITE_GTE_51:
                 # Spatialite >= 5.1 available, so we can try ST_ReducePrecision first,
                 # which should be faster.
+                # ST_ReducePrecision seems to crash on EMPTY geometry, so check
+                # ST_IsEmpty not being 0 (result can be -1, 0 or 1).
                 gridsize_op = f"""
-                    IIF(sub_gridsize.geom IS NULL,
+                    IIF(sub_gridsize.geom IS NULL OR ST_IsEmpty(sub_gridsize.geom) <> 0,
                         NULL,
                         IFNULL(
                             ST_ReducePrecision(sub_gridsize.geom, {gridsize}),
@@ -2737,12 +2747,11 @@ def _two_layer_vector_operation(
             logger.debug("Result was empty!")
 
         logger.info(f"Ready, took {datetime.now()-start_time}")
+        shutil.rmtree(tempdir, ignore_errors=True)
     except Exception:
         gfo.remove(output_path, missing_ok=True)
         gfo.remove(tmp_output_path, missing_ok=True)
         raise
-    finally:
-        shutil.rmtree(tempdir, ignore_errors=True)
 
 
 def calculate_two_layers(
@@ -3124,7 +3133,7 @@ def dissolve_singlethread(
     agg_columns: Optional[dict] = None,
     explodecollections: bool = False,
     gridsize: float = 0.0,
-    keep_empty_geoms: bool = True,
+    keep_empty_geoms: bool = False,
     where_post: Optional[str] = None,
     input_layer: Optional[str] = None,
     output_layer: Optional[str] = None,
@@ -3407,8 +3416,10 @@ def _format_apply_gridsize_operation(
         # It is not possible to return the original geometry if error stays after
         # makevalid, because spatialite functions return NULL for failures as well as
         # when the result is correctly NULL, so not possible to make the distinction.
+        # ST_ReducePrecision seems to crash on EMPTY geometry, so check ST_IsEmpty not
+        # being 0 (result can be -1, 0 or 1).
         gridsize_op = f"""
-            IIF({geometrycolumn} IS NULL,
+            IIF({geometrycolumn} IS NULL OR ST_IsEmpty({geometrycolumn}) <> 0,
                 NULL,
                 IFNULL(
                     ST_ReducePrecision({geometrycolumn}, {gridsize}),